requirements.txt
asgiref==3.2.10
autopep8==1.5.3
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
uuid==1.30
whitenoise==5.2.0
